feat: enable encryption/decryption of secrets
This pr also includes export capabilities for profiles for use with
rpc-go in a yaml file (with the contents encrypted).
The credential manager provided by the OS is used to store the key for encryption/decryption.
Depending on settings or OS, user may be prompted for OS credentials to launch console.
This PR also includes encryption/decryption of secrets in the database.

closes #398, #399, and #400

Co-authored-by: Matt Primrose <matt-primrose@users.noreply.github.com>